Introduced a DocumentController.
