+++ src/CognitiveSearch.UI/CognitiveSearch.UI/wwwroot/js/views/search-tables.js
@@ -35,7 +35,7 @@ Microsoft.Tables = {
                 resultsHtml += '<img alt="' + name + '" class="image-result" src="data:image/png;base64, ' + docresult.image.thumbnail_medium + '" title="' + docresult.metadata_storage_name + '" />';
             }
             else {
-                resultsHtml += '   <img alt="' + name + '" class="image-result" src="/api/search/getdocumentcoverimage?id=' + document_key + '" title="' + docresult.metadata_storage_name + '"onError="this.onerror=null;this.src=\'' + iconPath + '\';"/>';
+                resultsHtml += '   <img alt="' + name + '" class="image-result" src="/api/document/getcoverimage?id=' + document_key + '" title="' + docresult.metadata_storage_name + '"onError="this.onerror=null;this.src=\'' + iconPath + '\';"/>';
             }
 
             resultsHtml += '</a>';
